# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API DOCS",
        "description": "Advanced API documentation",
        "contact": {
            "email": "narushevich.maksim@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API documentation"
        }
    ],
    "paths": {
...
        "/products/{id}/reviews": {
            "get": {
                "tags": [
                    "Review"
                ],
                "summary": "Get all product's reviews",
                "description": "Returns all product's reviews details",
                "operationId": "getProductReviews",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Review"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Review"
                ],
                "summary": "Store a newly created resource in storage.",
                "description": "Create a new review.",
                "operationId": "addReview",
                "requestBody": {
                    "description": "Create review",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NewReview"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Null response"
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/products/{id}/reviews/{review}": {
            "get": {
                "tags": [
                    "Review"
                ],
                "summary": "Get review details",
                "description": "Returns review details",
                "operationId": "getReviewById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "review",
                        "in": "path",
                        "description": "ID of review to show",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Review"
                ],
                "summary": "Update the specified resource in storage.",
                "description": "Update review.",
                "operationId": "updateReview",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "review",
                        "in": "path",
                        "description": "ID of review to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update review",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NewReview"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Null response"
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Review"
                ],
                "summary": "Delete review",
                "description": "This can only be done by the logged in user.",
                "operationId": "deleteReview",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "review",
                        "in": "path",
                        "description": "ID of review to be deleted",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Review not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Product": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "stock": {
                        "type": "integer"
                    },
                    "price": {
                        "type": "integer"
                    },
                    "discount": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "Review": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "product_name": {
                        "type": "string"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "username": {
                        "type": "string"
                    },
                    "text": {
                        "type": "string"
                    },
                    "star": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "NewReview": {
                "type": "object",
                "allOf": [
                    {
                        "required": [
                            "star",
                            "text"
                        ],
                        "properties": {
                            "text": {
                                "type": "string"
                            },
                            "star": {
                                "type": "integer"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Oauth2 security",
                "name": "oauth2",
                "in": "header",
                "flows": {
                    "clientCredentials": {
                        "tokenUrl": "oauth/token",
                        "scopes": []
                    }
                },
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}